refactor: streamline data import and seed logic for users, roles, products, and categories

+++ src/database/seeders/20250313211101-users.js
@@ -1,8 +1,7 @@
 "use strict";
 /** @type {import('sequelize-cli').Migration} */
 
-const users = require("./data.json").users;
-const roles = require("./data.json").roles;
+const { users, roles, recoveries } = require("./data.json");
 
 module.exports = {
     async up(queryInterface, Sequelize) {
@@ -16,9 +15,9 @@ module.exports = {
          * }], {});
          */
 
-        await queryInterface.bulkInsert("Users", users, {});
-        await queryInterface.bulkInsert("Roles", roles, {});
-        // await queryInterface.bulkInsert("Recoveries", [], {});
+        if (users.length > 0) await queryInterface.bulkInsert("Users", users, {});
+        if (roles.length > 0) await queryInterface.bulkInsert("Roles", roles, {});
+        if (recoveries.length > 0) await queryInterface.bulkInsert("Recoveries", recoveries, {});
     },
 
     async down(queryInterface, Sequelize) {
@@ -29,8 +28,8 @@ module.exports = {
          * await queryInterface.bulkDelete('People', null, {});
          */
 
-        await queryInterface.bulkDelete("Users", null, {});
-        await queryInterface.bulkDelete("Roles", null, {});
-        await queryInterface.bulkDelete("Recoveries", null, {});
+        if (users.length > 0) await queryInterface.bulkDelete("Users", null, {});
+        if (roles.length > 0) await queryInterface.bulkDelete("Roles", null, {});
+        if (recoveries.length > 0) await queryInterface.bulkDelete("Recoveries", null, {});
     },
 };
